requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
python-dateutil>=2.8.2
pytz>=2023.3